fix: test-workflow-dryrun: avoid duplicate test names
test-workflow-dryrun.R loops over combinations of test_wd and
change_midstream values, generating the test_that() names.  Some of
these names collide:

    > test_df$TestName[duplicated(test_df$TestName)]
    [1] "basic workflow is correct using rbabylon.model_directory = 'model-examples'"
    [2] "basic workflow is correct using rbabylon.model_directory = 'model-examples'"
    [3] "basic workflow is correct using rbabylon.model_directory = 'model-examples' change_midstream"
    [4] "basic workflow is correct using rbabylon.model_directory = 'model-examples' change_midstream"

This causes issues with the upcoming mrgvalidate because it relies on
the test names being unique when a set of tests haven't been marked
with a test ID.

Add test_wd to the name (matching what's done elsewhere) so that the
test name is unique.